refactor(map): re-orginize into modules/hooks/components structure
- Implemented GeoJsonLayer component for rendering GeoJSON data on the map.
- Created MapBackButton for easy navigation back to the gallery.
- Added MapControls component for user geolocation and navigation controls.
- Developed MapInfoPanel to display information about markers and bounds.
- Introduced MapLoadingState to show loading indicators while fetching data.
- Built Mapbox component to integrate Mapbox GL with custom markers and popups.
- Created PhotoMarkerPin for displaying photo markers on the map.
- Developed PhotoPopup for showing detailed information about selected photos.
- Added PureMapbox component for a simplified Mapbox integration.
- Updated index files to export new components and hooks.
- Implemented hooks for managing map bounds, marker selection, and loading photo markers.
- Created utility functions for handling GPS coordinates and calculating map bounds.
- Built MapSection and MapboxContainer components to manage map state and interactions.
- Updated Explory page to utilize new map components and improve loading/error handling.
- Defined TypeScript types for map-related data structures.
